// client/src/vvm/pages/home/interfaces/interfaces.ts
export interface transformedData{
  petalLengthArray: number[];
  petalWidthArray: number[];
  sepalLengthArray: number[];
  sepalWidthArray: number[];
}

export interface imbdData{
  color: string;
  director_name: string;
  num_critic_for_reviews: number;
  duration: number;
  director_facebook_likes: number;
  actor_3_facebook_likes: number;
  actor_2_name: string;
  actor_1_facebook_likes: number;
  gross: number;
  genres: string;
  actor_1_name: string;
  movie_title: string;
  num_voted_users: number;
  cast_total_facebook_likes: number;
  actor_3_name: string;
  facenumber_in_poster: number;
  plot_keywords: string;
  movie_imdb_link: string;
  num_user_for_reviews: number;
  language: string;
  country: string;
  content_rating: string;
  budget: number;
  title_year: number;
  actor_2_facebook_likes: number;
  imdb_score: number;
  aspect_ratio: number;
  movie_facebook_likes: number;
}